PTR_ALIGNED_TO_SELF_SIZE: cast x to size_t rather than int to fix 64-bit build errors.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@15060 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/seqdb/seqdbgeneral.hpp
@@ -98,7 +98,7 @@ inline T SeqDB_GetBrokenUnaligned(const T * stdord_obj)
 #define ALIGNED_TO_POW2(x,y)  (! ((x) & (-y)))
 
 #define PTR_ALIGNED_TO_SELF_SIZE(x) \
-    (IS_POWER_OF_TWO(sizeof(*x)) && ALIGNED_TO_POW2(int(x), sizeof(*x)))
+    (IS_POWER_OF_TWO(sizeof(*x)) && ALIGNED_TO_POW2(size_t(x), sizeof(*x)))
 
 // Portable byte swapping from marshalled version
